+++ test/transform/resource/after-delombok/StandardExceptions.java
@@ -1,39 +1,36 @@
 class EmptyException extends Exception {
 	@java.lang.SuppressWarnings("all")
 	public EmptyException() {
+		this(null, null);
 	}
-
 	@java.lang.SuppressWarnings("all")
-	public EmptyException(final String message) {
+	public EmptyException(final java.lang.String message) {
+		this(message, null);
+	}
+	@java.lang.SuppressWarnings("all")
+	public EmptyException(final java.lang.Throwable cause) {
+		this(cause != null ? cause.getMessage() : null, cause);
+	}
+	@java.lang.SuppressWarnings("all")
+	public EmptyException(final java.lang.String message, final java.lang.Throwable cause) {
 		super(message);
-	}
-
-	@java.lang.SuppressWarnings("all")
-	public EmptyException(final Throwable cause) {
-		super(cause);
-	}
-
-	@java.lang.SuppressWarnings("all")
-	public EmptyException(final String message, final Throwable cause) {
-		super(message, cause);
+		if (cause != null) super.initCause(cause);
 	}
 }
 class NoArgsException extends Exception {
 	public NoArgsException() {
 	}
-
 	@java.lang.SuppressWarnings("all")
-	public NoArgsException(final String message) {
+	protected NoArgsException(final java.lang.String message) {
+		this(message, null);
+	}
+	@java.lang.SuppressWarnings("all")
+	protected NoArgsException(final java.lang.Throwable cause) {
+		this(cause != null ? cause.getMessage() : null, cause);
+	}
+	@java.lang.SuppressWarnings("all")
+	protected NoArgsException(final java.lang.String message, final java.lang.Throwable cause) {
 		super(message);
-	}
-
-	@java.lang.SuppressWarnings("all")
-	public NoArgsException(final Throwable cause) {
-		super(cause);
-	}
-
-	@java.lang.SuppressWarnings("all")
-	public NoArgsException(final String message, final Throwable cause) {
-		super(message, cause);
+		if (cause != null) super.initCause(cause);
 	}
 }