+++ test/transform/resource/before/StandardExceptions.java
@@ -1,8 +1,9 @@
+import lombok.AccessLevel;
 import lombok.experimental.StandardException;
 
 @StandardException class EmptyException extends Exception {
 }
-@StandardException class NoArgsException extends Exception {
+@StandardException(access = AccessLevel.PROTECTED) class NoArgsException extends Exception {
 	public NoArgsException() {
 	}
 }